src/pkg/exp/html/testdata/webkit/webkit02.dat

#data
<foo bar=qux/>
#errors
#document
| <html>
|   <head>
|   <body>
|     <foo>
|       bar="qux/"

#data
<p id="status"><noscript><strong>A</strong></noscript><span>B</span></p>
#errors
#document
| <html>
|   <head>
|   <body>
|     <p>
|       id="status"
|       <noscript>
|         "<strong>A</strong>"
|       <span>
|         "B"

#data
<div><sarcasm><div></div></sarcasm></div>
#errors
#document
| <html>
|   <head>
|   <body>
|     <div>
|       <sarcasm>
|         <div>

#data
<html><body><img src="" border="0" alt="><div>A</div></body></html>
#errors
#document
| <html>
|   <head>
|   <body>

#data
<table><td></tbody>A
#errors
#document
| <html>
|   <head>
|   <body>
|     "A"
|     <table>
|       <tbody>
|         <tr>
|           <td>

#data
<table><td></thead>A
#errors
#document
| <html>
|   <head>
|   <body>
|     <table>
|       <tbody>
|         <tr>
|           <td>
|             "A"

#data
<table><td></tfoot>A
#errors
#document
| <html>
|   <head>
|   <body>
|     <table>
|       <tbody>
|         <tr>
|           <td>
|             "A"

#data
<table><thead><td></tbody>A
#errors
#document
| <html>
|   <head>
|   <body>
|     <table>
|       <thead>
|         <tr>
|           <td>
|             "A"

#data
<legend>test</legend>
#errors
#document
| <html>
|   <head>
|   <body>
|     <legend>
|       "test"

#data
<table><input>
#errors
#document
| <html>
|   <head>
|   <body>
|     <input>
|     <table>

#data
<b><em><dcell><postfield><postfield><postfield><postfield><missing_glyph><missing_glyph><missing_glyph><missing_glyph><hkern><aside></b></em>
#errors
#document-fragment
div
#document
| <b>
|   <em>
|     <dcell>
|       <postfield>
|         <postfield>
|           <postfield>
|             <postfield>
|               <missing_glyph>
|                 <missing_glyph>
|                   <missing_glyph>
|                     <missing_glyph>
|                       <hkern>
| <aside>
|   <em>
|     <b>

#data
<isindex action="x">
#errors
#document-fragment
table
#document
| <form>
|   action="x"
| <hr>
| <label>
|   "This is a searchable index. Enter search keywords: "
|   <input>
|     name="isindex"
| <hr>

#data
<option><XH<optgroup></optgroup>
#errors
#document-fragment
select
#document
| <option>
